Use Select Condition is Split Position and Merge Positions
