+++ src/pages/MergePositions/Contents.tsx
@@ -4,9 +4,9 @@ import { Status } from 'util/types'
 
 import { Button } from 'components/buttons/Button'
 import { CenteredCard } from 'components/common/CenteredCard'
+import { SelectCondition } from 'components/form/SelectCondition'
 import { Amount } from 'components/mergePositions/Amount'
 import { MergePreview } from 'components/mergePositions/MergePreview'
-import { SelectCondition } from 'components/mergePositions/SelectCondition'
 import { SelectPosition } from 'components/mergePositions/SelectPosition'
 import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
 import { Row } from 'components/pureStyledComponents/Row'
